+++ morphac/controllers/base/binding/controllers.cc
@@ -1,10 +1,11 @@
 #include "pybind11/eigen.h"
 #include "pybind11/pybind11.h"
 
-#include "controllers/binding/include/controller_binding.h"
+#include "controllers/base/binding/include/controller_binding.h"
 
 namespace morphac {
 namespace controllers {
+namespace base {
 namespace binding {
 
 namespace py = pybind11;
@@ -14,5 +15,6 @@ PYBIND11_MODULE(_binding_controllers_python, m) {
 }
 
 }  // namespace binding
+}  // namespace base
 }  // namespace controllers
 }  // namespace morphac
