+++ morphac/controllers/base/binding/src/controller_binding.cc
@@ -1,12 +1,13 @@
-#include "controllers/binding/include/controller_binding.h"
+#include "controllers/base/binding/include/controller_binding.h"
 
 namespace morphac {
 namespace controllers {
+namespace base {
 namespace binding {
 
 namespace py = pybind11;
 
-using morphac::controllers::Controller;
+using morphac::controllers::base::Controller;
 
 void define_controller_binding(py::module& m) {
   py::class_<Controller, PyController> controller(m, "Controller");
@@ -15,5 +16,6 @@ void define_controller_binding(py::module& m) {
 }
 
 }  // binding
+}  // base
 }  // controllers
 }  // morphac
