+++ morphac/controllers/base/src/controller.cc
@@ -1,11 +1,13 @@
-#include "controllers/include/controller.h"
+#include "controllers/base/include/controller.h"
 
 namespace morphac {
 namespace controllers {
+namespace base {
 
 Controller::Controller() {}
 
 Controller::~Controller() {}
 
+} // namespace base
 } // namespace controllers
 } // namespace morphac
