+++ morphac/robot/CMakeLists.txt
@@ -23,4 +23,3 @@ install(FILES
   DESTINATION
   ${ROBOT_BINDING_INSTALL_DIR}
 )
-
